pyproject.toml
[project]
name = "conversation-system"
version = "1.0.0"
description = "Natural voice interactions with AI models using WhisperKit, Claude/Gemini, and ElevenLabs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.3.0",
    "python-dotenv>=1.1.0",
    "ruff-lsp>=0.0.62",
    "pygame>=2.5.0",
    "google-generativeai>=0.3.0",
    "structlog>=23.0.0",
    "click>=8.0.0",
]

[project.scripts]
conversation = "conversation_system.cli.main:cli"
